+++ app/widgets/Chat/chat.js
@@ -930,7 +930,7 @@ var Chat = {
     },
     appendMessage: function (idjidtime, data, prepend) {
         if (data.body === null) return;
-        console.log(data);
+
         var bubble = null,
             mergeMsg = false,
             msgStack,