+++ CHANGELOG.md
@@ -10,6 +10,7 @@ v0.22.6 (trunk)
 * Refactor the Stickers and File display and cleanup the related CSS
 * Complete the Manifest content to have a better PWA integration
 * Start the implementation of https://xmpp.org/extensions/xep-0444.html#disco-restricted for MUC
+* Add pagination on room presences list to prevent performances issues in large chatrooms
 
 v0.22.5
 ---------------------------